pyproject.toml
[tool.poetry]
name = "qabook"
version = "1.0.0"
license = "MIT"
description = "China science fiction"
authors = ["Bing D. Yee <bingdyee@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
dynaconf = "^3.2.6"
pymysql = "^1.1.1"
pymongo = "^4.9.1"
scrapy = "^2.11.2"
redis = "^5.1.0"
scrapy-redis = "^0.9.1"
numpy = "^2.1.1"
pandas = "^2.2.3"
streamlit = "^1.38.0"
matplotlib = "^3.9.2"
jupyter = "^1.1.1"
selenium = "^4.25.0"
faker = "^30.0.0"
openai = "^1.60.2"
langchain-ollama = "^0.2.2"
langchain-openai = "^0.3.2"
watchdog = "^6.0.0"
pypinyin = "^0.53.0"
langchain-community = "^0.3.21"
unstructured = "^0.17.2"
pypandoc = "^1.15"
faiss-cpu = "^1.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[tool.poetry.scripts]
qabook = "main:app"

[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
